requirements.txt
tiktokapipy>=0.2.0
requests>=2.31.0
python-dotenv>=1.0.0
playwright>=1.40.0
